backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
pytest
python-dotenv

